refactoring: some styling optimizations

+++ src/main/kotlin/com/ecwid/apiclient/v3/dto/profile/result/EmailLogoUploadResult.kt
@@ -3,5 +3,5 @@ package com.ecwid.apiclient.v3.dto.profile.result
 import com.ecwid.apiclient.v3.dto.common.ApiResultDTO
 
 data class EmailLogoUploadResult(
-	val id: Int = 0
+	val id: Long = 0
 ) : ApiResultDTO
